more feedback
also address some things the rebase missed
