+++ src/Microsoft.AspNetCore.SignalR.Common/Internal/Protocol/StreamItemMessage.cs
@@ -1,8 +1,5 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using Microsoft.Extensions.Internal;
 
 namespace Microsoft.AspNetCore.SignalR.Internal.Protocol
 {
